Refactor tests out into separate files
